pyproject.toml
[tool.poetry]
name = "Tavern Troupe Discord Bot"
version = "0.1.0"
description = ""
authors = ["Eitan Romanoff <eitan.romanoff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-twitter = "^3.5"
requests = "^2.25.0"
"discord.py" = "^1.5.1"
aiosqlite = "^0.16.0"
PyYAML = "^5.3.1"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
python-dateutil = "^2.8.1"
bs4 = "^0.0.1"
lxml = "^4.6.1"
libtmux = "^0.8.5"
aiofiles = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
